% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaggregate.R
\name{aggregate.challenge}
\alias{aggregate.challenge}
\title{Title}
\usage{
\method{aggregate}{challenge}(
  x,
  FUN = mean,
  na.treat,
  alpha = 0.05,
  p.adjust.method = "none",
  parallel = FALSE,
  progress = "none",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
