% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingMethods.R
\name{aggregateThenRank}
\alias{aggregateThenRank}
\title{Performs ranking via aggregate-then-rank}
\usage{
aggregateThenRank(object, FUN, ties.method = "min", ...)
}
\arguments{
\item{object}{The challenge object.}

\item{FUN}{The aggregation function, e.g. mean, median, min, max, function(x), quantile(x, probs=0.05).}

\item{ties.method}{A string specifying how ties are treated, see \code{\link[base:rank]{base::rank()}}.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{
An S3 object of class "ranked.list" to represent a ranked assessment data set.
}
\description{
Performs ranking by first aggregating performance values across all cases (e.g., with the mean, median or another quantile) for each algorithm.
This aggregate is then used to compute a rank for each algorithm.
}
\examples{

\dontrun{
 aggregateThenRank(challenge, FUN = mean, ties.method = "average", na.treat = 0)
}

}
\seealso{
Other ranking functions: 
\code{\link{rankThenAggregate}()},
\code{\link{testThenRank}()}
}
\concept{ranking functions}
