% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrap.R
\name{bootstrap.ranked.list}
\alias{bootstrap.ranked.list}
\title{Performs bootstrapping}
\usage{
\method{bootstrap}{ranked.list}(object, nboot, parallel = FALSE, progress = "text", ...)
}
\arguments{
\item{object}{The ranked assessment data set.}

\item{nboot}{The number of bootstrap samples.}

\item{parallel}{A boolean specifying whether parallel processing should be enabled.}

\item{progress}{A string specifying the type of progress indication.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{
An S3 object of class "bootstrap.list" to represent a bootstrapped, ranked assessment data set.
}
\description{
Performs bootstrapping on a ranked assessment data set and applies the ranking method to each bootstrap sample. One bootstrap sample of
a task with \code{n} cases consists of \code{n} cases randomly drawn with replacement from this task.
A total of \code{nboot} of these bootstrap samples are drawn.
}
\examples{

\dontrun{
 # perform bootstrapping with 1000 bootstrap samples using one CPU
 set.seed(1)
 ranking_bootstrapped <- bootstrap(ranking, nboot = 1000)
}

\dontrun{
 # perform bootstrapping using multiple CPUs (here: 8 CPUs)
 library(doParallel)
 registerDoParallel(cores=8)
 set.seed(1)
 ranking_bootstrapped <- bootstrap(ranking, nboot = 1000, parallel = TRUE, progress = "none")
 stopImplicitCluster()
}

}
