% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot.ranked.list}
\alias{boxplot.ranked.list}
\title{Creates dot- and boxplots}
\usage{
\method{boxplot}{ranked.list}(x, color = "blue", jitter.width = 0.25, ...)
}
\arguments{
\item{x}{The ranked assessment data set.}

\item{color}{A string specifying the color of the dots.}

\item{jitter.width}{A numeric value specifying the jitter width of the dots.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{

}
\description{
Creates dot- and boxplots visualizing the assessment data separately for each algorithm.
Boxplots representing descriptive statistics for all test cases (median, quartiles and outliers)
are combined with horizontally jittered dots representing individual test cases.
}
\examples{

}
\seealso{
\code{browseVignettes("challengeR")}

Other functions to visualize assessment data: 
\code{\link{podium.challenge}()},
\code{\link{podium.ranked.list}()},
\code{\link{rankingHeatmap.challenge}()},
\code{\link{rankingHeatmap.ranked.list}()}
}
\concept{functions to visualize assessment data}
