% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{consensus.ranked.list}
\alias{consensus.ranked.list}
\title{Computes a consensus ranking}
\usage{
\method{consensus}{ranked.list}(object, method, ...)
}
\arguments{
\item{object}{The ranked asssessment data set.}

\item{method}{A string specifying the method to derive the consensus ranking, see \code{\link[relations:consensus]{relations::consensus()}} for the methods. Consensus ranking according to mean ranks across tasks if method="euclidean" where in case of ties (equal ranks for multiple algorithms) the average rank is used, i.e. ties.method="average".}

\item{...}{Further arguments passed to or from other functions.}
}
\value{

}
\description{
Computes a consensus ranking (rank aggregation) across tasks.
}
