% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendrogram.R
\name{dendrogram.ranked.list}
\alias{dendrogram.ranked.list}
\title{Creates a cluster dendrogram}
\usage{
\method{dendrogram}{ranked.list}(object, dist = "symdiff", method = "complete", ...)
}
\arguments{
\item{object}{The ranked assessment data set.}

\item{dist}{A string specifying the distance measure to be used, see \code{\link[relations:dissimilarity]{relations::dissimilarity()}}.}

\item{method}{A string specifying agglomeration method to be used, see \code{\link[stats:hclust]{stats::hclust()}}.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{

}
\description{
Creates a cluster dendrogram from a ranked assessment data set.
}
\examples{

}
\seealso{
\code{browseVignettes("challengeR")}

Other functions to visualize cross-task insights: 
\code{\link{stability.ranked.list}()},
\code{\link{stabilityByAlgorithm.bootstrap.list}()},
\code{\link{stabilityByTask.bootstrap.list}()}
}
\concept{functions to visualize cross-task insights}
