% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsplot.R
\name{methodsplot.challenge}
\alias{methodsplot.challenge}
\title{Creates line plots}
\usage{
\method{methodsplot}{challenge}(
  x,
  na.treat = NULL,
  methods = list(testBased = . \%>\% test() \%>\% rank(ties.method = "min"),
    meanThenRank = . \%>\% aggregate(FUN = "mean") \%>\% rank(ties.method = "min"),
    medianThenRank = . \%>\% aggregate(FUN = "median") \%>\% rank(ties.method = "min"),
    rankThenMean = . \%>\% rank(ties.method = "min") \%>\% aggregate(FUN = "mean") \%>\%
    rank(ties.method = "min"), rankThenMedian = . \%>\% rank(ties.method = "min") \%>\%
    aggregate(FUN = "median") \%>\% rank(ties.method = "min")),
  ordering,
  ...
)
}
\arguments{
\item{x}{The challenge object.}

\item{na.treat}{Indicates how missing perfomance values are treated if sanity check is enabled. It can be 'na.rm', numeric value or function.
For a numeric value or function, NAs will be replaced by the specified values. For 'na.rm', rows that contain missing values will be removed.}

\item{methods}{A list of ranking methods that should be incorporated.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{

}
\description{
Create line plots that visualize the robustness of ranking across different ranking methods from a challenge object.
}
\examples{

}
\seealso{
\code{browseVignettes("challengeR")}

Other functions to visualize ranking stability: 
\code{\link{significanceMap.ranked.list}()},
\code{\link{stabilityByTask.bootstrap.list}()},
\code{\link{violin.bootstrap.list}()}
}
\concept{functions to visualize ranking stability}
