% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podium.R
\name{podium.challenge}
\alias{podium.challenge}
\title{Creates a podium plot}
\usage{
\method{podium}{challenge}(
  object,
  ordering,
  xlab = NULL,
  ylab = NULL,
  lines.show = FALSE,
  lines.alpha = 0.2,
  lines.lwd = 1,
  lines.lty = 1,
  col,
  lines.col = col,
  dots.pch = 19,
  dots.cex = 1,
  places.lty = 2,
  places.col = 1,
  legendfn = function(algs, cols) {     legend("topleft", algs, lwd = 1, col = cols, bg
    = "white") },
  layout.heights = c(1, 0.4),
  ...
)
}
\arguments{
\item{object}{The challenge object.}

\item{xlab}{A string specifying the x-axis label.}

\item{ylab}{A string specifying the y-axis label.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{

}
\description{
Creates a podium plot from a challenge object.
}
\examples{

}
\seealso{
\code{browseVignettes("challengeR")}

Other functions to visualize assessment data: 
\code{\link{boxplot.ranked.list}()},
\code{\link{podium.ranked.list}()},
\code{\link{rankingHeatmap.challenge}()},
\code{\link{rankingHeatmap.ranked.list}()}
}
\concept{functions to visualize assessment data}
