% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podium.R
\name{podium.ranked.list}
\alias{podium.ranked.list}
\title{Creates podium plots}
\usage{
\method{podium}{ranked.list}(
  object,
  xlab = "Podium",
  ylab = "Performance",
  lines.show = TRUE,
  lines.alpha = 0.2,
  lines.lwd = 1,
  lines.lty = 1,
  lines.col = col,
  col,
  dots.pch = 19,
  dots.cex = 1,
  places.lty = 2,
  places.col = 1,
  legendfn = function(algs, cols) {     legend("topright", algs, lwd = 1, col = cols,
    bg = "white") },
  layout.heights = c(1, 0.4),
  ...
)
}
\arguments{
\item{object}{The ranked asssessment data set.}

\item{xlab}{A string specifying the x-axis label.}

\item{ylab}{A string specifying the y-axis label.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{

}
\description{
Creates podium plots from one or more ranked assessment data sets.
}
\examples{

}
\seealso{
\code{browseVignettes("challengeR")}

Other functions to visualize assessment data: 
\code{\link{boxplot.ranked.list}()},
\code{\link{podium.challenge}()},
\code{\link{rankingHeatmap.challenge}()},
\code{\link{rankingHeatmap.ranked.list}()}
}
\concept{functions to visualize assessment data}
