% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingHeatmap.R
\name{rankingHeatmap.challenge}
\alias{rankingHeatmap.challenge}
\title{Creates a ranking heatmap}
\usage{
\method{rankingHeatmap}{challenge}(x, ordering, ties.method = "min", ...)
}
\arguments{
\item{x}{The challenge object.}

\item{ties.method}{A string specifying how ties are treated, see \code{\link[base:rank]{base::rank()}}.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{

}
\description{
Creates a ranking heatmap from a challenge object.
}
\examples{

}
\seealso{
\code{browseVignettes("challengeR")}

Other functions to visualize assessment data: 
\code{\link{boxplot.ranked.list}()},
\code{\link{podium.challenge}()},
\code{\link{podium.ranked.list}()},
\code{\link{rankingHeatmap.ranked.list}()}
}
\concept{functions to visualize assessment data}
