% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingHeatmap.R
\name{rankingHeatmap.ranked.list}
\alias{rankingHeatmap.ranked.list}
\title{Creates ranking heatmaps}
\usage{
\method{rankingHeatmap}{ranked.list}(x, ties.method = "min", ...)
}
\arguments{
\item{x}{The ranked asssessment data set.}

\item{ties.method}{A string specifying how ties are treated, see \code{\link[base:rank]{base::rank()}}.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{

}
\description{
Creates ranking heatmaps from one or more ranked assessment data sets.
}
\examples{

}
\seealso{
\code{browseVignettes("challengeR")}

Other functions to visualize assessment data: 
\code{\link{boxplot.ranked.list}()},
\code{\link{podium.challenge}()},
\code{\link{podium.ranked.list}()},
\code{\link{rankingHeatmap.challenge}()}
}
\concept{functions to visualize assessment data}
