% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingMethods.R
\name{rankThenAggregate}
\alias{rankThenAggregate}
\title{Performs ranking via rank-then-aggregate}
\usage{
rankThenAggregate(object, FUN, ties.method = "min")
}
\arguments{
\item{object}{The challenge object.}

\item{FUN}{The aggregation function, e.g., mean, median, min, max, function(x), quantile(x, probs=0.05).}

\item{ties.method}{A string specifying how ties are treated, see \code{\link[base:rank]{base::rank()}}.}
}
\value{
An S3 object of class "ranked.list" to represent a ranked assessment data set.
}
\description{
Performs ranking by first computing a rank for each case for each algorithm ("rank first").
The final rank is based on the aggregated ranks for the cases. This ranking method handles missing values implicitly
by assigning the worst rank to missing algorithm performances.
}
\examples{
\dontrun{
 rankThenAggregate(challenge, FUN = mean)
}

}
\seealso{
Other ranking functions: 
\code{\link{aggregateThenRank}()},
\code{\link{testThenRank}()}
}
\concept{ranking functions}
