% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report.bootstrap.list}
\alias{report.bootstrap.list}
\title{Generates a benchmarking report with bootstrapping results}
\usage{
\method{report}{bootstrap.list}(
  object,
  consensus,
  file,
  title = "<Challenge name>",
  colors = default_colors,
  format = "PDF",
  latex_engine = "pdflatex",
  clean = TRUE,
  fig.format = NULL,
  dpi = 150,
  open = TRUE,
  ...
)
}
\arguments{
\item{object}{The ranked (bootstrapped) assessment data set.}

\item{consensus}{The rank aggregation across tasks (consensus ranking). Only needed for a multi-task data set.}

\item{file}{A string specifying the file name of the report. It allows for specifying the output file path as well,
otherwise the working directory is used. If \code{file} does not have a file extension, an extension will be automatically
added according to the output format given in \code{format}. If the argument is omitted, the report is created in a
temporary folder with file name "report".}

\item{title}{A string specifying the title of the report.}

\item{colors}{The color scheme that is applied to the plots.}

\item{format}{A string specifying the format of the report. The options are "PDF", "HTML" or "Word".}

\item{latex_engine}{A string specifying the LaTeX engine for producing PDF output. The Options are "pdflatex", "lualatex", and "xelatex".}

\item{clean}{A boolean indicating whether intermediate files (e.g. individual plots) should be kept. Using \code{TRUE} will clean
intermediate files that are created during rendering.}

\item{fig.format}{A vector of strings containing the file format of the figures that are not removed if \code{clean} is set to \code{FALSE}.
The options are "jpeg", "png" and "pdf", e.g. \code{fig.format = c("jpeg", "png", "pdf")}.}

\item{dpi}{A positive integer specifying the resolution of the generated plot (\code{fig.format} "jpeg" or "png") in dots per inch (DPI).}

\item{open}{A boolean specifying whether the report should be opened with the default system viewer after generation.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{

}
\description{
Generates a benchmarking report in PDF, HTML or Word format with bootstrapping results.
It contains the rankings, plots of the raw assessment data and plots of the ranking stability.
For multi-task challenges, it also contains plots of cross-task insights. If you are interested in
the individual plots as separate files, set argument \code{clean} to \code{FALSE} and specify \code{fig.format}.
}
