% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significanceMap.R
\name{significanceMap.ranked.list}
\alias{significanceMap.ranked.list}
\title{Creates significance maps}
\usage{
\method{significanceMap}{ranked.list}(
  object,
  alpha = 0.05,
  p.adjust.method = "holm",
  order = FALSE,
  size.rank = 0.3 * theme_get()$text$size,
  ...
)
}
\arguments{
\item{object}{The ranked assessment data set.}

\item{alpha}{A numeric values specifying the significance level.}

\item{p.adjust.method}{A string specifying the adjustment method for multiple testing, see \code{\link[stats:p.adjust]{stats::p.adjust()}}.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{

}
\description{
Creates significance maps from a ranked assessment data set.
}
\examples{

}
\seealso{
\code{browseVignettes("challengeR")}

Other functions to visualize ranking stability: 
\code{\link{methodsplot.challenge}()},
\code{\link{stabilityByTask.bootstrap.list}()},
\code{\link{violin.bootstrap.list}()}
}
\concept{functions to visualize ranking stability}
