% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{stability.ranked.list}
\alias{stability.ranked.list}
\title{Creates a blob plot across tasks}
\usage{
\method{stability}{ranked.list}(
  x,
  ordering,
  probs = c(0.025, 0.975),
  max_size = 6,
  freq = FALSE,
  shape = 4,
  ...
)
}
\arguments{
\item{x}{The ranked asssessment data set.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{

}
\description{
Creates a blob plots visualizing the ranking variability across tasks.
}
\examples{

}
\seealso{
\code{browseVignettes("challengeR")}

Other functions to visualize cross-task insights: 
\code{\link{dendrogram.ranked.list}()},
\code{\link{stabilityByAlgorithm.bootstrap.list}()},
\code{\link{stabilityByTask.bootstrap.list}()}
}
\concept{functions to visualize cross-task insights}
