% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{stabilityByAlgorithm.bootstrap.list}
\alias{stabilityByAlgorithm.bootstrap.list}
\title{Creates blob plots or stacked frequency plots stratified by algorithm}
\usage{
\method{stabilityByAlgorithm}{bootstrap.list}(
  x,
  ordering,
  stacked = FALSE,
  probs = c(0.025, 0.975),
  max_size = 3,
  shape = 4,
  freq = FALSE,
  single = FALSE,
  ...
)
}
\arguments{
\item{x}{The bootstrapped, ranked assessment data set.}

\item{stacked}{A boolean specifying whether a stacked frequency plot (\code{stacked = TRUE}) or blob plot (\code{stacked = FALSE}) should be created.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{

}
\description{
Creates blob plots (\code{stacked = FALSE}) or stacked frequency plots (\code{stacked = TRUE}) for each algorithm
from a bootstrapped, ranked assessment data set.
}
\examples{

}
\seealso{
\code{browseVignettes("challengeR")}

Other functions to visualize cross-task insights: 
\code{\link{dendrogram.ranked.list}()},
\code{\link{stability.ranked.list}()},
\code{\link{stabilityByTask.bootstrap.list}()}
}
\concept{functions to visualize cross-task insights}
