% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset.ranked.list}
\alias{subset.ranked.list}
\title{Extracts a subset of algorithms or tasks}
\usage{
\method{subset}{ranked.list}(x, top, tasks, ...)
}
\arguments{
\item{x}{The ranked asssessment data set.}

\item{top}{A positive integer specifying the amount of top performing algorithms to be retrieved.}

\item{tasks}{A vector of strings containing the task identifiers that should remain in the subset.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{
An S3 object of class "ranked.list" to represent a ranked assessment data set.
}
\description{
Extracts the top performing algorithms or a subset of tasks.
}
\section{Reports for subsets (top list) of algorithms}{

If ties are present in the ranking, the subset will consist of more than \code{top} algorithms.
Line plots for ranking robustness can be used to check whether algorithms performing well in other
ranking methods are excluded. Bootstrapping still takes entire uncertainty into account.
Podium plots and ranking heatmaps neglect excluded algorithms. Only available for single-task challenges
(for multi-task challenges not sensible because each task would contain a different set of algorithms).
}

\section{Reports for subsets of tasks}{

You may want to recompute the consensus ranking after creating the subset. An error will be raised
if a task identifier is not contained in the assessment data set to avoid subsequent errors.
}

\examples{

\dontrun{
 # only show the top 3 algorithms according to the chosen ranking method
 subset(ranking, top = 3) \%>\% report(...)
}

\dontrun{
 # restrict report to tasks "task1" and "task2"
 subset(ranking, tasks=c("task1", "task2")) \%>\% report(...)
}

}
