% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingMethods.R
\name{testThenRank}
\alias{testThenRank}
\title{Performs ranking via test-then-rank}
\usage{
testThenRank(object, ties.method = "min", ...)
}
\arguments{
\item{object}{The challenge object.}

\item{ties.method}{A string specifying how ties are treated, see \code{\link[base:rank]{base::rank()}}.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{
An S3 object of class "ranked.list" to represent a ranked assessment data set.
}
\description{
Computes statistical hypothesis tests based on Wilcoxon signed rank test for each possible
pair of algorithms to assess differences in metric values between the algorithms.
Then ranking is performed according to the number of significant one-sided test results.
If algorithms have the same number of significant test results, then they obtain the same rank.
}
\examples{
\dontrun{
 testThenRank(challenge,
              alpha=0.05, # significance level
              p.adjust.method="none", # method for adjustment for multiple testing, see ?p.adjust
              na.treat = 0)
}

}
\seealso{
Other ranking functions: 
\code{\link{aggregateThenRank}()},
\code{\link{rankThenAggregate}()}
}
\concept{ranking functions}
