% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin.R
\name{violin.bootstrap.list}
\alias{violin.bootstrap.list}
\title{Creates a violin plot}
\usage{
\method{violin}{bootstrap.list}(x, ...)
}
\arguments{
\item{x}{The bootstrapped, ranked assessment data set.}

\item{...}{Further arguments passed to or from other functions.}
}
\value{

}
\description{
Creates a violin plot from a bootstrapped, ranked assessment data set.
}
\examples{

}
\seealso{
\code{browseVignettes("challengeR")}

Other functions to visualize ranking stability: 
\code{\link{methodsplot.challenge}()},
\code{\link{significanceMap.ranked.list}()},
\code{\link{stabilityByTask.bootstrap.list}()}
}
\concept{functions to visualize ranking stability}
